reg_time: switch from int to unsigned long for time type
This will ease the work afterward because the arduino function working
with time have "unsigned long" as type

+++ arduino/register/reg_time.c
@@ -1,11 +1,11 @@
 #include <stdbool.h>
 
 // time in microseconds
-static int time = 0;
+static unsigned long time = 0;
 // incrementation of the time after each loop
 static int incr = 0;
 
-int reg_get_time() {
+unsigned long reg_get_time() {
     return time;
 }
 
@@ -13,7 +13,7 @@ void reg_set_time_interval(int microseconds) {
     incr = microseconds;
 }
 
-int reg_increment_time() {
+unsigned long reg_increment_time() {
     time += incr;
     return time;
 }
